reconstructionCuda performs regularised superresolution similar to reconstruction.cc it is implemented in CUDA and therefore a bit faster than the TBB parallized version (speedup is currently ~40 compared to single CPU). Beware that after the first build TBB might give you a warning about the compiler. It vanishes after the second build and has no effect. TODO full Integration into IRTK CMake environment -> need to build separately for now TODO optimizations

+++ image++/src/irtkGenericImage.cc
@@ -391,7 +391,7 @@ template <class VoxelType> VoxelType irtkGenericImage<VoxelType>::GetSDPad(doubl
 
   if (m > 0) std /= double(m);
 
-  return sqrt(std);
+  return sqrt((double)std);
 }
 
 template <class VoxelType> void irtkGenericImage<VoxelType>::GetMaxPosition(irtkPoint& p, int ds, int) const
